// src/app/Models/Models.ts
  export interface Rank {
    tier: string
    leagueId: string
    queue: string
    name: string
    entries: RankSummoner[]
  }
  
  export interface RankSummoner {
    summonerId: string
    summonerName: string
    leaguePoints: number
    rank: string
    wins: number
    losses: number
    veteran: boolean
    inactive: boolean
    freshBlood: boolean
    hotStreak: boolean
  }